# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Metis API",
    "description": "Metis API documentation",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:4000",
  "basePath": "/",
...
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/v1/api/reportUser": {
      "post": {
        "tags": ["Report user"],
        "description": "Report a user",
        "parameters": [
          {
            "name": "signup",
            "in": "body",
            "description": "Create new account in system",
            "schema": {
              "$ref": "#/definitions/ReportUserRequest"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Report User success",
            "schema": {
              "$ref": "#/definitions/DefaultResponse"
            }
          },
          "403": {
            "description": "Not authenticated",
            "schema": {
              "$ref": "#/definitions/DefaultResponse"
            }
          }
        }
      }
    },
...
    "/v1/api/data/messages": {
      "post": {
        "tags": ["Messages"],
        "description": "Create new channel message in blockchain",
        "parameters": [
          {
            "name": "message creation",
            "in": "body",
            "description": "User sends a message",
            "schema": {
              "$ref": "#/definitions/MessageCreationRequest"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/DefaultResponse"
            }
          },
          "403": {
            "description": "Not authenticated",
            "schema": {
              "$ref": "#/definitions/DefaultResponse"
            }
          }
        }
      }
    },
...
    "/v1/api/channels/invite": {
      "post": {
        "tags": ["Invites"],
        "description": "Send an invite",
        "parameters": [
          {
            "name": "message creation",
            "in": "body",
            "description": "User sends a message",
            "schema": {
              "$ref": "#/definitions/InviteCreationRequest"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/DefaultResponse"
            }
          },
          "403": {
            "description": "Not authenticated",
            "schema": {
              "$ref": "#/definitions/DefaultResponse"
            }
          }
        }
      }
    },
    "/v1/api/channels/invites": {
      "get": {
        "tags": ["Invites"],
        "summary": "Get all user invites",
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/InviteResponse"
            }
          },
          "403": {
            "description": "Not authenticated",
            "schema": {
              "$ref": "#/definitions/DefaultResponse"
            }
          }
        }
      }
    },
    "/v1/api/channels/import": {
      "post": {
        "tags": ["Invites"],
        "summary": "Accept invites",
        "description": "Accept invites",
        "parameters": [
          {
            "name": "message creation",
            "in": "body",
            "description": "User sends a message",
            "schema": {
              "$ref": "#/definitions/AcceptInviteRequest"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/DefaultResponse"
            }
          },
          "403": {
            "description": "Not authenticated",
            "schema": {
              "$ref": "#/definitions/DefaultResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "ReportUserRequest": {
      "properties": {
        "reporter": {
          "type": "string",
          "description": "user Jup account"
        },
        "message": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    },
...
    "MessageData": {
      "type": "object",
      "properties": {
        "channel": {
          "type": "object",
          "$ref": "#/definitions/Channel"
        },
        "message": {
          "type": "string"
        },
        "mentions": {
          "type": "array",
          "description": "String array of mentioned members",
          "items": {}
        },
        "replyMessage": {
          "type": "string"
        },
        "replyRecipientName": {
          "type": "string"
        },
        "isInvitation": {
          "type": "boolean"
        },
        "messageVersion": {
          "type": "string"
        }
      }
    },
    "ChannelRecord": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "passphrase": {
          "type": "string"
        },
        "account": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "publicKey": {
          "type": "string"
        },
        "sender": {
          "type": "string"
        },
        "createdBy": {
          "type": "string"
        },
        "date": {
          "type": "number"
        },
        "confirmed": {
          "type": "boolean"
        }
      }
    },
    "InviteCreationRequest": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "recipient": {
              "type": "string",
              "description": "Alias or Jup id recipient"
            },
            "channel": {
              "type": "object",
              "$ref": "#/definitions/Channel"
            }
          }
        }
      }
    },
    "MessageCreationRequest": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "description": "Message to send",
          "$ref": "#/definitions/MessageData"
        },
        "tableData": {
          "type": "object",
          "description": "Channel info",
          "$ref": "#/definitions/ChannelRecord"
        },
        "channel": {
          "type": "object",
          "description": "Channel info",
          "$ref": "#/definitions/Channel"
        },
        "mentions": {
          "type": "array",
          "items": {}
        }
      }
    },
...
    "Channel": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "channel_record": {
          "type": "object",
          "$ref": "#/definitions/ChannelRecord"
        },
        "date": {
          "type": "number"
        },
        "token": {
          "type": "string"
        }
      }
    },
    "InviteData": {
      "type": "object",
      "properties": {
        "recipient": {
          "type": "string"
        },
        "sender": {
          "type": "string"
        },
        "channel": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
            },
            "channel_record": {
              "type": "object",
              "$ref": "#/definitions/ChannelRecord"
            },
            "date": {
              "type": "number"
            },
            "name": {
              "type": "string"
            }
          }
        },
        "dataType": {
          "type": "string"
        }
      }
    },
    "AcceptInviteRequest": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
            },
            "passphrase": {
              "type": "string"
            },
            "account": {
              "type": "string"
            },
            "password": {
              "type": "string"
            },
            "name": {
              "type": "string"
            },
            "publicKey": {
              "type": "string"
            },
            "sender": {
              "type": "string"
            },
            "createdBy": {
              "type": "string"
            },
            "date": {
              "type": "number"
            },
            "confirmed": {
              "type": "boolean"
            },
            "invited": {
              "type": "boolean"
            }
          }
        }
      }
    },
    "InviteResponse": {
      "type": "object",
      "properties": {
        "recordsFound": {
          "type": "number"
        },
        "pending": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/InviteData"
          }
        },
        "records": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/InviteData"
          }
        },
        "last_record": {
          "type": "object",
          "$ref": "#/definitions/InviteData"
        }
      }
    },
...
    "DefaultResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message":{
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "jwt": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
  "security": [
    {
      "jwt": []
    }
  ]
}